Fix potential overflow in calculateCompactSize

+++ presto-spi/src/main/java/com/facebook/presto/spi/block/DictionaryBlock.java
@@ -203,7 +203,7 @@ public class DictionaryBlock
 
     private void calculateCompactSize()
     {
-        int sizeInBytes = 0;
+        long sizeInBytes = 0;
         int uniqueIds = 0;
         boolean[] seen = new boolean[dictionary.getPositionCount()];
         for (int i = 0; i < positionCount; i++) {
